+++ packages/client/mutations/fragments/CompleteTaskFrag.ts
@@ -4,6 +4,7 @@ graphql`
     id
     content
     plaintextContent
+    title
     createdAt
     createdBy
     createdByUser {
@@ -21,8 +22,7 @@ graphql`
     sortOrder
     status
     tags
-    threadId
-    threadSource
+    discussionId
     threadSortOrder
     threadParentId
     updatedAt