+++ packages/server/graphql/intranetSchema/intranetSchema.ts
@@ -17,6 +17,7 @@ import enableSAMLForDomain from './mutations/enableSAMLForDomain'
 import endOldMeetings from './mutations/endOldMeetings'
 import flagConversionModal from './mutations/flagConversionModal'
 import flagOverLimit from './mutations/flagOverLimit'
+import lockTeams from './mutations/lockTeams'
 import loginSAML from './mutations/loginSAML'
 import messageAllSlackUsers from './mutations/messageAllSlackUsers'
 import profileCPU from './mutations/profileCPU'
@@ -38,6 +39,7 @@ import pingActionTick from './queries/pingActionTick'
 import signups from './queries/signups'
 import user from './queries/user'
 import users from './queries/users'
+import updateEmail from './mutations/updateEmail'
 
 const query = new GraphQLObjectType<any, GQLContext>({
   name: 'Query',
@@ -73,6 +75,7 @@ const mutation = new GraphQLObjectType<any, GQLContext>({
       endOldMeetings,
       flagConversionModal,
       flagOverLimit,
+      lockTeams,
       loginSAML,
       enableSAMLForDomain,
       messageAllSlackUsers,
@@ -86,7 +89,8 @@ const mutation = new GraphQLObjectType<any, GQLContext>({
       stripeSucceedPayment,
       stripeUpdateCreditCard,
       stripeUpdateInvoiceItem,
-      stripeInvoiceFinalized
+      stripeInvoiceFinalized,
+      updateEmail
     } as any)
 })