+++ packages/server/graphql/intranetSchema/mutations/checkRethinkPgEquality.ts
@@ -1,6 +1,5 @@
-import {GraphQLNonNull, GraphQLString} from 'graphql'
+import {GraphQLBoolean, GraphQLInt, GraphQLNonNull, GraphQLString} from 'graphql'
 import {requireSU} from '../../../utils/authorization'
-import standardError from '../../../utils/standardError'
 import fs from 'fs'
 import path from 'path'
 import checkUserEq from '../../../postgres/utils/checkUserEq'
@@ -9,10 +8,20 @@ import checkTeamEq from '../../../postgres/utils/checkTeamEq'
 const tableResolvers = {
   User: checkUserEq,
   Team: checkTeamEq
-} as {[key: string]: () => Promise<{[key: string]: any}>}
+} as {
+  [key: string]: (
+    pageSize?: number,
+    startPage?: number,
+    slice?: boolean
+  ) => Promise<{[key: string]: any}>
+}
 
-const checkEqAndWriteOutput = async (tableName: string, fileLocation: string): Promise<void> => {
-  const errors = await tableResolvers[tableName]()
+const checkEqAndWriteOutput = async (
+  tableName: string,
+  fileLocation: string,
+  maxErrors = 10
+): Promise<void> => {
+  const errors = await tableResolvers[tableName](maxErrors)
   await fs.promises.writeFile(fileLocation, JSON.stringify(errors))
 }
 
@@ -23,20 +32,32 @@ const checkRethinkPgEquality = {
     tableName: {
       type: GraphQLNonNull(GraphQLString),
       description: 'The table name to be compared'
+    },
+    maxErrors: {
+      type: GraphQLInt,
+      default: 10,
+      description: 'How many errors should be returned'
+    },
+    writeToFile: {
+      type: GraphQLBoolean,
+      default: false,
+      description: 'Whether the output should be written to file'
     }
   },
-  resolve: async (_source, {tableName}, {authToken}) => {
+  resolve: async (_source, {tableName, maxErrors, writeToFile}, {authToken}) => {
     // AUTH
     requireSU(authToken)
 
     // VALIDATION
     if (!tableResolvers.hasOwnProperty(tableName)) {
-      return standardError(
-        new Error(`That table name either doesn't exist or hasn't yet been implemented.`)
-      )
+      return `That table name either doesn't exist or hasn't yet been implemented.`
     }
 
     // RESOLUTION
+    if (!writeToFile) {
+      const errors = await tableResolvers[tableName](maxErrors)
+      return JSON.stringify(errors)
+    }
     const fileName = `${tableName}-${new Date()}`
     const fileLocation = path.join(
       process.cwd(),
@@ -45,7 +66,7 @@ const checkRethinkPgEquality = {
       fileName
     )
     await fs.promises.mkdir(path.dirname(fileLocation), {recursive: true})
-    checkEqAndWriteOutput(tableName, fileLocation)
+    checkEqAndWriteOutput(tableName, fileLocation, maxErrors)
     return `Please check ${fileLocation} for output results, it will appear in a few mins.`
   }
 }