+++ packages/server/graphql/intranetSchema/mutations/flagOverLimit.ts
@@ -4,6 +4,7 @@ import {requireSU} from '../../../utils/authorization'
 import {GQLContext} from '../../graphql'
 import FlagOverLimitPayload from '../../types/FlagOverLimitPayload'
 import updateUser from '../../../postgres/queries/updateUser'
+import {USER_OVERLIMIT_COPY_LIMIT} from '../../../postgres/constants'
 
 const flagOverLimit = {
   type: FlagOverLimitPayload,
@@ -25,15 +26,19 @@ const flagOverLimit = {
 
     // VALIDATION
     const organizationUsers = await dataLoader.get('organizationUsersByOrgId').load(orgId)
-
     if (organizationUsers.length === 0) return {error: {message: 'OrgId has no members'}}
+
+    if (copy.length > USER_OVERLIMIT_COPY_LIMIT) {
+      return {error: {message: `copy must be ${USER_OVERLIMIT_COPY_LIMIT} chars or less`}}
+    }
 
     // RESOLUTION
     const userIds = organizationUsers.map(({userId}) => userId)
-    await Promise.all([
-      updateUser({overLimitCopy: copy}, userIds),
-      db.writeMany('User', userIds, {overLimitCopy: copy || null})
-    ])
+    const updates = {
+      overLimitCopy: copy === null ? '' : copy,
+      updatedAt: new Date()
+    }
+    await Promise.all([updateUser(updates, userIds), db.writeMany('User', userIds, updates)])
     return {userIds}
   }
 }