+++ packages/server/graphql/mutations/removeAgendaItem.ts
@@ -1,12 +1,12 @@
 import {GraphQLID, GraphQLNonNull} from 'graphql'
-import getRethink from '../../database/rethinkDriver'
-import RemoveAgendaItemPayload from '../types/RemoveAgendaItemPayload'
-import publish from '../../utils/publish'
-import {getUserId, isTeamMember} from '../../utils/authorization'
-import standardError from '../../utils/standardError'
-import removeStagesFromMeetings from './helpers/removeStagesFromMeetings'
-import AgendaItemsStage from '../../database/types/AgendaItemsStage'
 import {SubscriptionChannel} from 'parabol-client/types/constEnums'
+import getRethink from '../../database/rethinkDriver'
+import AgendaItemsStage from '../../database/types/AgendaItemsStage'
+import {getUserId, isTeamMember} from '../../utils/authorization'
+import publish from '../../utils/publish'
+import standardError from '../../utils/standardError'
+import RemoveAgendaItemPayload from '../types/RemoveAgendaItemPayload'
+import removeStagesFromMeetings from './helpers/removeStagesFromMeetings'
 
 export default {
   type: RemoveAgendaItemPayload,
@@ -34,17 +34,18 @@ export default {
     const agendaItem = await r
       .table('AgendaItem')
       .get(agendaItemId)
-      .delete({returnChanges: true})('changes')(0)('old_val')
+      .update(
+        {isActive: false},
+        {returnChanges: true}
+      )('changes')(0)('old_val')
       .default(null)
       .run()
     if (!agendaItem) {
       return standardError(new Error('Agenda item not found'), {userId: viewerId})
     }
     const filterFn = (stage: AgendaItemsStage) => stage.agendaItemId === agendaItemId
-    const meetingIds = await removeStagesFromMeetings(filterFn, teamId, dataLoader)
-    // safe to do so because we guarantee only 1 action meeting at the same time
-    const [meetingId] = meetingIds
-    const data = {agendaItem, meetingId}
+    await removeStagesFromMeetings(filterFn, teamId, dataLoader)
+    const data = {agendaItem, meetingId: agendaItem.meetingId}
     publish(SubscriptionChannel.TEAM, teamId, 'RemoveAgendaItemPayload', data, subOptions)
     return data
   }