+++ packages/server/graphql/mutations/startRetrospective.ts
@@ -6,6 +6,8 @@ import MeetingRetrospective from '../../database/types/MeetingRetrospective'
 import MeetingSettingsRetrospective from '../../database/types/MeetingSettingsRetrospective'
 import Organization from '../../database/types/Organization'
 import RetroMeetingMember from '../../database/types/RetroMeetingMember'
+import generateUID from '../../generateUID'
+import updateTeamByTeamId from '../../postgres/queries/updateTeamByTeamId'
 import {getUserId, isTeamMember} from '../../utils/authorization'
 import publish from '../../utils/publish'
 import standardError from '../../utils/standardError'
@@ -14,7 +16,6 @@ import StartRetrospectivePayload from '../types/StartRetrospectivePayload'
 import createNewMeetingPhases from './helpers/createNewMeetingPhases'
 import {startSlackMeeting} from './helpers/notifySlack'
 import sendMeetingStartToSegment from './helpers/sendMeetingStartToSegment'
-import updateTeamByTeamId from '../../postgres/queries/updateTeamByTeamId'
 
 export default {
   type: new GraphQLNonNull(StartRetrospectivePayload),
@@ -40,7 +41,7 @@ export default {
       return standardError(new Error('Team not found'), {userId: viewerId})
     }
 
-    const meetingType = 'retrospective' as MeetingTypeEnum
+    const meetingType: MeetingTypeEnum = 'retrospective'
 
     // RESOLUTION
     const meetingCount = await r
@@ -51,9 +52,11 @@ export default {
       .default(0)
       .run()
 
+    const meetingId = generateUID()
     const phases = await createNewMeetingPhases(
       viewerId,
       teamId,
+      meetingId,
       meetingCount,
       meetingType,
       dataLoader
@@ -70,6 +73,7 @@ export default {
       .load({teamId, meetingType})) as MeetingSettingsRetrospective
     const {totalVotes, maxVotesPerGroup, selectedTemplateId} = meetingSettings
     const meeting = new MeetingRetrospective({
+      id: meetingId,
       teamId,
       meetingCount,
       phases,
@@ -80,7 +84,6 @@ export default {
       templateId: selectedTemplateId
     })
 
-    const meetingId = meeting.id
     const template = await dataLoader.get('meetingTemplates').load(selectedTemplateId)
     await r
       .table('NewMeeting')
@@ -103,6 +106,10 @@ export default {
       return {error: {message: 'Meeting already started'}}
     }
 
+    const updates = {
+      lastMeetingType: meetingType,
+      updatedAt: new Date()
+    }
     await Promise.all([
       r
         .table('MeetingMember')
@@ -113,9 +120,9 @@ export default {
       r
         .table('Team')
         .get(teamId)
-        .update({lastMeetingType: meetingType})
+        .update(updates)
         .run(),
-      updateTeamByTeamId({lastMeetingType: meetingType}, teamId)
+      updateTeamByTeamId(updates, teamId)
     ])
 
     startSlackMeeting(meetingId, teamId, dataLoader).catch(console.log)