+++ packages/server/postgres/utils/checkEqBase.ts
@@ -1,21 +1,78 @@
 import {RTable, TableSchema} from '../../database/stricterR'
+import areEqual from 'fbjs/lib/areEqual'
+import getPg from '../getPg'
 
-type WithId<T> = T & {id: string}
+interface DBDoc {
+  id: string
+  [key: string]: any
+}
+interface RethinkDoc extends DBDoc {}
+interface PGDoc extends DBDoc {}
 
-export type CustomResolver = (rethinkValue: string, pgValue: string) => boolean
-interface IError<rethinkType, pgType> {
-  [key: string]: {
-    error: string | string[]
-    rethinkRow?: Partial<rethinkType>
-    pgRow?: Partial<pgType>
-  }
+interface IError {
+  [key: string]:
+    | {
+        error: string[]
+        rethinkRow: Partial<RethinkDoc>
+        pgRow: Partial<PGDoc>
+      }
+    | number
+    | boolean
+    | undefined
+    | string[]
+  rethinkRecordsCompared: number
+  foundErrors: number
+  missingPgRows: string[]
+  rethinkRowCount?: number
+  pgRowCount?: number
+  areRowCountsEqual?: boolean
 }
 
-function addNeFieldsToErrors<rethinkType, pgType>(
-  errors: IError<rethinkType, pgType>,
+function getPairNeFields(
+  rethinkRow: RethinkDoc,
+  pgRow: PGDoc,
+  alwaysDefinedFields: string[],
+  maybeUndefinedFieldsDefaultValues: {[key: string]: any},
+  maybeNullFieldsDefaultValues?: {[key: string]: any}
+): string[] {
+  const neFields = [] as string[]
+
+  for (const f of alwaysDefinedFields) {
+    const [rethinkValue, pgValue] = [rethinkRow[f], pgRow[f]]
+    if (!areEqual(rethinkValue, pgValue)) {
+      neFields.push(f)
+    }
+  }
+  for (const [maybeUndefinedField, defaultValueForUndefinedField] of Object.entries(
+    maybeUndefinedFieldsDefaultValues
+  )) {
+    const [rethinkValue, pgValue] = [rethinkRow[maybeUndefinedField], pgRow[maybeUndefinedField]]
+
+    if (rethinkValue === undefined) {
+      if (areEqual(pgValue, defaultValueForUndefinedField)) {
+        continue
+      }
+    } else if (rethinkValue === null && pgValue !== null) {
+      if (
+        maybeNullFieldsDefaultValues &&
+        areEqual(maybeNullFieldsDefaultValues[maybeUndefinedField], pgValue)
+      ) {
+        continue
+      }
+    } else if (areEqual(pgValue, rethinkValue)) {
+      continue
+    }
+
+    neFields.push(maybeUndefinedField)
+  }
+  return neFields
+}
+
+function addNeFieldsToErrors(
+  errors: IError,
   neFields: string[],
-  rethinkRow: rethinkType,
-  pgRow: pgType,
+  rethinkRow: RethinkDoc,
+  pgRow: PGDoc,
   rowId: string
 ): void {
   const rethinkNeFields = {}
@@ -31,47 +88,80 @@ function addNeFieldsToErrors<rethinkType, pgType>(
   }
 }
 
-export async function checkTableEq<rethinkType, pgType>(
+const getPgRowCount = async (tableName: string) => {
+  const pg = getPg()
+  const queryRes = await pg.query(`
+    SELECT COUNT(*) FROM "${tableName}";
+  `)
+  return Number(queryRes.rows[0].count)
+}
+
+export async function checkTableEq(
+  tableName: string,
   rethinkQuery: RTable<TableSchema>,
-  pgQuery: (ids: string[]) => Promise<pgType[]>,
-  getPairNeFieldsCb: (rethinkRow: rethinkType, pgRow: pgType) => string[]
-): Promise<IError<rethinkType, pgType>> {
-  const errors = {} as IError<rethinkType, pgType>
+  pgQuery: (ids: string[]) => Promise<PGDoc[] | null>,
+  alwaysDefinedFields: string[],
+  maybeUndefinedFieldsDefaultValues: {[key: string]: any},
+  maybeNullFieldsDefaultValues: {[key: string]: any},
+  maxErrors = 10
+): Promise<IError> {
+  const errors: IError = {
+    rethinkRecordsCompared: 0,
+    foundErrors: 0,
+    missingPgRows: []
+  }
   const batchSize = 3000
 
   for (let i = 0; i < 1e5; i++) {
-    console.log(i)
+    if (errors.foundErrors >= maxErrors) {
+      return errors
+    }
+
     const offset = batchSize * i
     const rethinkRows = (await rethinkQuery
       .skip(offset)
       .limit(batchSize)
-      .run()) as rethinkType[]
+      .run()) as RethinkDoc[]
     if (!rethinkRows.length) {
       break
     }
 
-    const ids = rethinkRows.map((t) => (t as WithId<rethinkType>).id) as string[]
-    const pgRows = await pgQuery(ids)
-    const pgRowsById = {} as {[key: string]: pgType}
+    const ids = rethinkRows.map((t) => t.id)
+    const pgRows = (await pgQuery(ids)) ?? []
+    const pgRowsById = {} as {[key: string]: PGDoc}
     pgRows.forEach((pgRow) => {
-      pgRowsById[(pgRow as WithId<pgType>).id] = pgRow
+      pgRowsById[pgRow.id] = pgRow
     })
 
     for (const rethinkRow of rethinkRows) {
-      const id = (rethinkRow as WithId<rethinkType>).id
+      if (errors.foundErrors >= maxErrors) {
+        return errors
+      }
+      errors.rethinkRecordsCompared += 1
+      const id = rethinkRow.id
       const pgRow = pgRowsById[id]
+
       if (!pgRow) {
-        errors[id] = {
-          error: 'No pg row found for rethink row'
-        }
+        errors.missingPgRows.push(id)
+        errors.foundErrors += 1
         continue
       }
-      const neFields = getPairNeFieldsCb(rethinkRow, pgRow)
+      const neFields = getPairNeFields(
+        rethinkRow,
+        pgRow,
+        alwaysDefinedFields,
+        maybeUndefinedFieldsDefaultValues,
+        maybeNullFieldsDefaultValues
+      )
+
       if (neFields.length) {
-        addNeFieldsToErrors<rethinkType, pgType>(errors, neFields, rethinkRow, pgRow, id)
+        addNeFieldsToErrors(errors, neFields, rethinkRow, pgRow, id)
+        errors.foundErrors += 1
       }
     }
   }
-
+  errors.rethinkRowCount = errors.rethinkRecordsCompared
+  errors.pgRowCount = await getPgRowCount(tableName)
+  errors.areRowCountsEqual = errors.rethinkRowCount === errors.pgRowCount
   return errors
 }